CR fixes #2; add docstrings, explanatory comments, move wrap_user to apps/users
