+++ zapisy/apps/theses/users.py
@@ -2,7 +2,6 @@
 on users of the theses system"""
 from enum import Enum
 
-from django.http import Http404
 from django.contrib.auth.models import User
 from apps.users.models import BaseUser, Employee, Student, is_user_in_group
 
@@ -56,25 +55,6 @@ def get_user_type(base_user: BaseUser) -> ThesisUserType:
     elif isinstance(base_user, Student):
         return ThesisUserType.STUDENT
     return ThesisUserType.NONE
-
-
-def wrap_user(user: User) -> BaseUser:
-    """Given an instance of contrib.auth.models.User,
-    wrap it into an instance of BaseUser.
-
-    This is necessary because of legacy logic present in apps.users.models:
-    all users are instances of Student or Employee (both subclasses of BaseUser),
-    which link to their corresponding instance of Django's User with a OneToOneField.
-    This is most likely legacy logic from the days when Django didn't offer the possibility
-    to change the user auth model; if this is refactored,
-    this function will not be necessary anymore.
-    """
-
-    if BaseUser.is_employee(user):
-        return user.employee
-    elif BaseUser.is_student(user):
-        return user.student
-    raise Http404("invalid user")
 
 
 def get_theses_user_full_name(user: BaseUser):
